Refactor: Move role and user-roles models to roles module
This commit moves the Role and UserRoles models from the users module to the roles module. The imports have been updated accordingly in the related files. The models have been
