+++ src/app.module.ts
@@ -4,8 +4,8 @@ import { UsersModule } from './users/users.module'
 import { ConfigModule } from '@nestjs/config'
 import { User } from './users/model/users.model'
 import { RolesModule } from './roles/roles.module'
-import { Role } from './roles/roles.model'
-import { UserRoles } from './roles/user-roles.model'
+import { Role } from './roles/model/roles.model'
+import { UserRoles } from './roles/model/user-roles.model'
 import { AuthModule } from './auth/auth.module'
 import { PostsModule } from './posts/posts.module'
 import { Post } from './posts/posts.model'
