+++ src/roles/roles.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Get, Param, Post } from '@nestjs/common'
 import { RolesService } from './roles.service'
 import { CreateRoleDto } from './dto/create-role.dto'
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger'
-import { Role } from './roles.model'
+import { Role } from './model/roles.model'
 
 @ApiTags('Роли')
 @Controller('roles')
