+++ src/roles/roles.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common'
 import { CreateRoleDto } from './dto/create-role.dto'
 import { InjectModel } from '@nestjs/sequelize'
-import { Role } from './roles.model'
+import { Role } from './model/roles.model'
 
 @Injectable()
 export class RolesService {
