+++ src/users/model/users.model.ts
@@ -9,8 +9,8 @@ import {
     Table,
 } from 'sequelize-typescript'
 import { Post } from 'src/posts/posts.model'
-import { Role } from 'src/roles/roles.model'
-import { UserRoles } from 'src/roles/user-roles.model'
+import { Role } from 'src/roles/model/roles.model'
+import { UserRoles } from 'src/roles/model/user-roles.model'
 
 interface UserCreationAttrs {
     email: string
